# backend/docs/swagger.yaml
basePath: /
definitions:
...
  cards.hnAmount:
    properties:
      currencyCode:
        type: string
      value:
        type: integer
    type: object
  cards.hnAuthRequest:
    properties:
      authorizationRequestId:
        type: string
      cardId:
        type: string
      id:
        type: string
      merchantDetails:
        $ref: '#/definitions/cards.hnMerchant'
      transactionAmount:
        $ref: '#/definitions/cards.hnAmount'
      type:
        type: string
    type: object
  cards.hnAuthResponse:
    properties:
      approvedTransactionAmount:
        $ref: '#/definitions/cards.hnAmount'
      authorizationResponseCode:
        type: string
    type: object
  cards.hnMerchant:
    properties:
      categoryCode:
        type: string
      name:
        type: string
    type: object
  cards.issueCardRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      member_id:
        type: string
    required:
    - email
    - first_name
    - last_name
    - member_id
    type: object
  cards.issueCardResponse:
    properties:
      card_id:
        type: string
      card_token:
        type: string
      cardholder_id:
        type: string
    type: object
  cards.loadWalletRequest:
    properties:
      amount_cents:
        type: integer
      member_id:
        type: string
    required:
    - amount_cents
    - member_id
    type: object
  cards.loadWalletResponse:
    properties:
      new_balance_cents:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend for Tally — a group card-spending app with JIT authorization
    and a double-entry ledger.
  title: Tally API
  version: "1.0"
paths:
...
  /v1/cards/issue:
    post:
      consumes:
      - application/json
      description: Creates a Highnote cardholder and virtual card for a member. The
        returned card_token is used to identify the card in JIT authorization webhooks.
      parameters:
      - description: Member identity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cards.issueCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cards.issueCardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Issue a virtual card
      tags:
      - cards
...
  /v1/wallets/load:
    post:
      consumes:
      - application/json
      description: Credits a member's Tally wallet balance (Tier 1 funding source).
        Also syncs the load to Highnote on a best-effort basis.
      parameters:
      - description: Load details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cards.loadWalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cards.loadWalletResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Load a member's wallet
      tags:
      - cards
  /v1/webhooks/highnote/authorization:
    post:
      consumes:
      - application/json
      description: Receives Highnote card authorization events and runs the 5-tier
        funding waterfall (Tally balance → primary bank → secondary bank → leader
        overwrite → partial auth). Requires X-Tally-Signature and Idempotency-Key
        headers.
      parameters:
      - description: 'HMAC-SHA256 signature: sha256=<hex>'
        in: header
        name: X-Tally-Signature
        required: true
        type: string
      - description: Unique key to deduplicate retries
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Highnote authorization event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cards.hnAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cards.hnAuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Highnote JIT authorization webhook
      tags:
      - webhooks
swagger: "2.0"